+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_models.py
@@ -6,6 +6,7 @@
 
 from ._generated.models._models import LanguageInput
 from ._generated.models._models import MultiLanguageInput
+from enum import Enum
 
 
 class DictMixin(object):
@@ -57,6 +58,10 @@ class DictMixin(object):
         if key in self.__dict__:
             return self.__dict__[key]
         return default
+
+class Encoding(str, Enum):
+    """The encoding in which the offsets will be returned in"""
+    grapheme = "grapheme"
 
 
 class DetectedLanguage(DictMixin):
@@ -168,12 +173,12 @@ class CategorizedEntity(DictMixin):
     :type category: str
     :param subcategory: Entity subcategory, such as Age/Year/TimeRange etc
     :type subcategory: str
-    :param grapheme_offset: Start position (in Unicode characters) for the
-        entity text.
-    :type grapheme_offset: int
-    :param grapheme_length: Length (in Unicode characters) for the entity
-        text.
-    :type grapheme_length: int
+    :param offset: Start position for the entity text. Its encoding is
+        determined by the value for `encoding` passed into `recognize_entities`
+    :type offset: int
+    :param length: Length for the entity text. Its encoding is
+        determined by the value for `encoding` passed into `recognize_entities`
+    :type length: int
     :param confidence_score: Confidence score between 0 and 1 of the extracted
         entity.
     :type confidence_score: float
@@ -183,8 +188,8 @@ class CategorizedEntity(DictMixin):
         self.text = kwargs.get('text', None)
         self.category = kwargs.get('category', None)
         self.subcategory = kwargs.get('subcategory', None)
-        self.grapheme_offset = kwargs.get('grapheme_offset', None)
-        self.grapheme_length = kwargs.get('grapheme_length', None)
+        self.offset = kwargs.get('offset', None)
+        self.length = kwargs.get('length', None)
         self.confidence_score = kwargs.get('confidence_score', None)
 
     @classmethod
@@ -193,15 +198,15 @@ class CategorizedEntity(DictMixin):
             text=entity.text,
             category=entity.category,
             subcategory=entity.subcategory,
-            grapheme_offset=entity.offset,
-            grapheme_length=entity.length,
+            offset=entity.offset,
+            length=entity.length,
             confidence_score=entity.confidence_score,
         )
 
     def __repr__(self):
-        return "CategorizedEntity(text={}, category={}, subcategory={}, grapheme_offset={}, grapheme_length={}, " \
-               "confidence_score={})".format(self.text, self.category, self.subcategory, self.grapheme_offset,
-                                  self.grapheme_length, self.confidence_score)[:1024]
+        return "CategorizedEntity(text={}, category={}, subcategory={}, offset={}, length={}, " \
+               "confidence_score={})".format(self.text, self.category, self.subcategory, self.offset,
+                                  self.length, self.confidence_score)[:1024]
 
 
 class TextAnalyticsError(DictMixin):
@@ -391,16 +396,17 @@ class TextDocumentStatistics(DictMixin):
     """TextDocumentStatistics contains information about
     the document payload.
 
-    :param grapheme_count: Number of text elements recognized in
-        the document.
-    :type grapheme_count: int
+    :param count: Number of text elements recognized in
+        the document. Its encoding is determined by the value for `encoding` passed into
+        the endpoint.
+    :type count: int
     :param transaction_count: Number of transactions for the
         document.
     :type transaction_count: int
     """
 
     def __init__(self, **kwargs):
-        self.grapheme_count = kwargs.get("grapheme_count", None)
+        self.count = kwargs.get("count", None)
         self.transaction_count = kwargs.get("transaction_count", None)
 
     @classmethod
@@ -408,13 +414,13 @@ class TextDocumentStatistics(DictMixin):
         if stats is None:
             return None
         return cls(
-            grapheme_count=stats.characters_count,
+            count=stats.characters_count,
             transaction_count=stats.transactions_count,
         )
 
     def __repr__(self):
-        return "TextDocumentStatistics(grapheme_count={}, transaction_count={})" \
-            .format(self.grapheme_count, self.transaction_count)[:1024]
+        return "TextDocumentStatistics(count={}, transaction_count={})" \
+            .format(self.count, self.transaction_count)[:1024]
 
 
 class DocumentError(DictMixin):
@@ -549,32 +555,32 @@ class LinkedEntityMatch(DictMixin):
     :type confidence_score: float
     :param text: Entity text as appears in the request.
     :type text: str
-    :param grapheme_offset: Start position (in Unicode characters) for the
-        entity match text.
-    :type grapheme_offset: int
-    :param grapheme_length: Length (in Unicode characters) for the entity
-        match text.
-    :type grapheme_length: int
+    :param offset: Start position for the entity match text. Its encoding is
+        determined by the value for `encoding` passed into `recognize_linked_entities`
+    :type offset: int
+    :param length: Length for the entity match text. Its encoding is
+        determined by the value for `encoding` passed into `recognize_linked_entities`
+    :type length: int
     """
 
     def __init__(self, **kwargs):
         self.confidence_score = kwargs.get("confidence_score", None)
         self.text = kwargs.get("text", None)
-        self.grapheme_offset = kwargs.get("grapheme_offset", None)
-        self.grapheme_length = kwargs.get("grapheme_length", None)
+        self.offset = kwargs.get("offset", None)
+        self.length = kwargs.get("length", None)
 
     @classmethod
     def _from_generated(cls, match):
         return cls(
             confidence_score=match.confidence_score,
             text=match.text,
-            grapheme_offset=match.offset,
-            grapheme_length=match.length
+            offset=match.offset,
+            length=match.length
         )
 
     def __repr__(self):
-        return "LinkedEntityMatch(confidence_score={}, text={}, grapheme_offset={}, grapheme_length={})" \
-            .format(self.confidence_score, self.text, self.grapheme_offset, self.grapheme_length)[:1024]
+        return "LinkedEntityMatch(confidence_score={}, text={}, offset={}, length={})" \
+            .format(self.confidence_score, self.text, self.offset, self.length)[:1024]
 
 
 class TextDocumentInput(MultiLanguageInput):
@@ -654,19 +660,22 @@ class SentenceSentiment(DictMixin):
         and 1 for the sentence for all labels.
     :type confidence_scores:
         ~azure.ai.textanalytics.SentimentConfidenceScores
-    :param grapheme_offset: The sentence offset from the start of the
-        document.
-    :type grapheme_offset: int
-    :param grapheme_length: The length of the sentence by Unicode standard.
-    :type grapheme_length: int
+    :param offset: The sentence offset from the start of the
+        document. Its encoding is determined by the value for `encoding`
+        passed into `analyze_sentiment`
+    :type offset: int
+    :param length: The length of the sentence by Unicode standard.
+        Its encoding is determined by the value for `encoding`
+        passed into `analyze_sentiment`
+    :type length: int
     """
 
     def __init__(self, **kwargs):
         self.text = kwargs.get("text", None)
         self.sentiment = kwargs.get("sentiment", None)
         self.confidence_scores = kwargs.get("confidence_scores", None)
-        self.grapheme_offset = kwargs.get("grapheme_offset", None)
-        self.grapheme_length = kwargs.get("grapheme_length", None)
+        self.offset = kwargs.get("offset", None)
+        self.length = kwargs.get("length", None)
 
     @classmethod
     def _from_generated(cls, sentence):
@@ -674,14 +683,14 @@ class SentenceSentiment(DictMixin):
             text=sentence.text,
             sentiment=sentence.sentiment,
             confidence_scores=SentimentConfidenceScores._from_generated(sentence.confidence_scores),  # pylint: disable=protected-access
-            grapheme_offset=sentence.offset,
-            grapheme_length=sentence.length
+            offset=sentence.offset,
+            length=sentence.length
         )
 
     def __repr__(self):
-        return "SentenceSentiment(text={}, sentiment={}, confidence_scores={}, grapheme_offset={}, "\
-            "grapheme_length={})".format(self.text, self.sentiment, repr(self.confidence_scores),
-            self.grapheme_offset, self.grapheme_length
+        return "SentenceSentiment(text={}, sentiment={}, confidence_scores={}, offset={}, "\
+            "length={})".format(self.text, self.sentiment, repr(self.confidence_scores),
+            self.offset, self.length
         )[:1024]