+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_recognize_entities.py
@@ -44,8 +44,8 @@ class TestRecognizeEntities(TextAnalyticsTest):
             for entity in doc.entities:
                 self.assertIsNotNone(entity.text)
                 self.assertIsNotNone(entity.category)
-                self.assertIsNotNone(entity.grapheme_offset)
-                self.assertIsNotNone(entity.grapheme_length)
+                self.assertIsNotNone(entity.offset)
+                self.assertIsNotNone(entity.length)
                 self.assertIsNotNone(entity.confidence_score)
 
     @GlobalTextAnalyticsAccountPreparer()
@@ -63,8 +63,8 @@ class TestRecognizeEntities(TextAnalyticsTest):
             for entity in doc.entities:
                 self.assertIsNotNone(entity.text)
                 self.assertIsNotNone(entity.category)
-                self.assertIsNotNone(entity.grapheme_offset)
-                self.assertIsNotNone(entity.grapheme_length)
+                self.assertIsNotNone(entity.offset)
+                self.assertIsNotNone(entity.length)
                 self.assertIsNotNone(entity.confidence_score)
 
     @GlobalTextAnalyticsAccountPreparer()